# rubocop: reduce every regex evaluation for performance