--- a/lib/fluent/config/literal_parser.rb
+++ b/lib/fluent/config/literal_parser.rb
@@ -98,11 +98,11 @@ module Fluent
             else
               return string.join
             end
-          elsif check(/[^"]#{LINE_END_WITHOUT_SPACING_AND_COMMENT}/)
-            if s = check(/[^\\]#{LINE_END_WITHOUT_SPACING_AND_COMMENT}/)
+          elsif check(/[^"]#{LINE_END_WITHOUT_SPACING_AND_COMMENT}/o)
+            if s = check(/[^\\]#{LINE_END_WITHOUT_SPACING_AND_COMMENT}/o)
               string << s
             end
-            skip(/[^"]#{LINE_END_WITHOUT_SPACING_AND_COMMENT}/)
+            skip(/[^"]#{LINE_END_WITHOUT_SPACING_AND_COMMENT}/o)
           elsif s = scan(/\\./)
             string << eval_escape_char(s[1,1])
           elsif skip(/\#\{/)